+++ src/Redux/rockets/rockets.js
@@ -48,16 +48,4 @@ const rocketsReducer = (state = initialState, action) => {
   }
 };
 
-const url = 'https://api.spacexdata.com/v3/rockets';
-export const getRockets = () => async (dispatch) => {
-  dispatch(getRocketsRequest());
-  try {
-    const response = await fetch(url);
-    const data = await response.json();
-    dispatch(getRocketsSuccess(data));
-  } catch (err) {
-    dispatch(getRocketsFailure(err.message));
-  }
-};
-
 export default rocketsReducer;
